feature(pause): refactoring protocols
